requirements.txt
coverage==4.5.1
Django==2.2.10
django-nose==1.4.6
djangorestframework==3.10.2 
nose==1.3.7 
psycopg2==2.8.6
django-extensions==3.2.3